DT-4725 Refactored apis and updated auth service

+++ src/landing-page/landing-page.controller.ts
@@ -8,12 +8,15 @@ import {
   Query,
   Delete,
   HttpException,
+  Req,
+  BadRequestException,
 } from '@nestjs/common';
 import { LandingPageService } from './landing-page.service';
 import { UsersService } from 'src/users/users.service';
 import { LandingPagePublishRepository } from './landing-page-publish.repository';
 import { LeadsFilterDto } from './dto/leads-dto';
 import { Protected } from 'src/auth/auth.decorator';
+import { AuthService } from 'src/auth/auth.service';
 
 @Controller({
   version: '1',
@@ -26,6 +29,7 @@ export class LandingPageController {
     private readonly landingPageService: LandingPageService,
     private readonly landingPagePublishRepository: LandingPagePublishRepository,
     private readonly usersService: UsersService,
+    private readonly authService: AuthService,
   ) {}
 
   @Get('mapped-domain')
@@ -108,17 +112,22 @@ export class LandingPageController {
 
   @Protected()
   @Delete('leads/:id')
-  async deleteLead(@Param('id') id: number, @Query('slug') slug: string, @Query('leadType') leadType: string) {
+  async deleteLead(@Param('id') id: number, @Query('slug') slug: string, @Query('leadType') leadType: string, @Req() req) {
     try {
       const brand = await this.usersService.getBrandIdBySlug(slug);
       if (!brand) {
         throw new Error(`Brand not found for slug: ${slug}`);
       }
+      if (!this.authService.validateUser(brand.id, req.user)) {
+        throw new BadRequestException(
+          `Unauthorized to access resources for ${slug}`,
+        );
+      }
       const response = await this.landingPageService.deleteLead(id, brand.id, leadType);
 
       return response;
     } catch (error) {
-     throw new HttpException('Failed to delete lead', error?.status || 500);
+     throw new HttpException(error?.message || 'Failed to delete lead', error?.status || 500);
     }
   }
   @Get('leads/export')
@@ -155,8 +164,14 @@ export class LandingPageController {
   async createOrUpdate(
     @Param('brandId') brandId: number,
     @Body() createLandingPageDto: any,
+    @Req() req
   ) {
     try {
+      if (!this.authService.validateUser(brandId, req.user)) {
+        throw new BadRequestException(
+          `Unauthorized to access resources for ${brandId}`,
+        );
+      }
       const data = await this.landingPageService.createOrUpdate(
         brandId,
         createLandingPageDto,
@@ -175,18 +190,23 @@ export class LandingPageController {
   @Post('publish/:slug')
   async createOrUpdatePublish(
     @Param('slug') slug: string,
-    @Body()
-    publishDto: {
+    @Body() publishDto: {
       publishStatus: boolean;
       domainType: string;
       domain?: string;
       customDomainStatus?: string;
     },
+    @Req() req,
   ) {
     try {
       const brand = await this.usersService.getBrandIdBySlug(slug);
       if (!brand) {
         throw new Error(`Brand not found for slug: ${slug}`);
+      }
+      if(!this.authService.validateUser(brand.id, req.user)) {
+        throw new BadRequestException(
+          `Unauthorized to access resources for ${slug}`,
+        );
       }
       const data = await this.landingPageService.createOrUpdatePublish(
         brand?.id,
@@ -256,11 +276,17 @@ export class LandingPageController {
     @Param('slug') slug: string,
     @Param('sectionSlug') sectionSlug: string,
     @Body() createLandingPageDto: any,
+    @Req() req
   ) {
     try {
       const brand = await this.usersService.getBrandIdBySlug(slug);
       if (!brand) {
         throw new Error(`Brand not found for slug: ${slug}`);
+      }
+      if(!this.authService.validateUser(brand.id, req.user)) {
+        throw new BadRequestException(
+          `Unauthorized to access resources for ${slug}`,
+        );
       }
       const data = await this.landingPageService.createOrUpdateSection(
         brand?.id,
